refactored delegate mthod and scene
